#7 code refactoring
