+++ app/src/main/java/com/example/githubrepowatcher/domain/usecases/SessionUseCase.kt
@@ -6,9 +6,8 @@ import com.example.githubrepowatcher.domain.repository.SessionRepository
 class SessionUseCase(
     private val sessionRepository: SessionRepository
 ) {
-    fun saveAuthToken(keyValueStorage: KeyValueStorage) = sessionRepository.saveAuthToken(
-        keyValueStorage
-    )
+    fun saveAuthToken(keyValueStorage: KeyValueStorage) =
+        sessionRepository.saveAuthToken(keyValueStorage)
 
     fun getAuthToken(): KeyValueStorage? = sessionRepository.getAuthToken()
