+++ app/src/main/java/com/example/githubrepowatcher/presentation/fragments/AuthFragment.kt
@@ -10,7 +10,7 @@ import android.view.ViewGroup
 import android.widget.Toast
 import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.ViewModelProvider
+import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.lifecycleScope
 import com.example.githubrepowatcher.R
 import com.example.githubrepowatcher.databinding.FragmentAuthBinding
@@ -25,7 +25,7 @@ class AuthFragment : Fragment() {
     private val binding: FragmentAuthBinding
         get() = _binding ?: throw RuntimeException("FragmentAuthBinding is null")
 
-    private lateinit var sessionViewModel: SessionViewModel
+    private val sessionViewModel: SessionViewModel by activityViewModels()
 
     private val authToken = "asd123"
 
@@ -40,7 +40,6 @@ class AuthFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        sessionViewModel = ViewModelProvider(this)[SessionViewModel::class.java]
 
         if (sessionViewModel.getAuthToken() != null) {
             Toast.makeText(requireContext(), "Success", Toast.LENGTH_SHORT).show()
