+++ app/src/main/res/layout/fragment_auth.xml
@@ -40,6 +40,7 @@
                 android:paddingTop="4dp"
                 android:paddingBottom="16dp"
                 app:boxStrokeColor="@color/input_token"
+                app:hintTextColor="@color/input_token"
                 android:hint="@string/token_hint"
                 app:errorEnabled="false"
                 app:hintTextAppearance="@style/InputTextAppearance"
